Ensure stopped services don't get restarted on accident during updates

+++ scripts/update/btsync.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 if [[ -f /install/.btsync.lock ]]; then
   if [[ ! -f /etc/systemd/system/resilio-sync.service ]]; then
-    systemctl stop resilo-sync
+    active=$(systemctl is-active lounge)
+    if [[ $active == "active" ]]; then
+      systemctl stop resilo-sync
+    fi
     MASTER=$(cut -d: -f1 < /root/.master.info)
     BTSYNCIP=$(ip route get 1 | sed -n 's/^.*src \([0-9.]*\) .*$/\1/p')
     cat > /etc/resilio-sync/config.json <<RSCONF
@@ -22,6 +25,8 @@ RSCONF
     sed -i "s/rslsync:rslsync/${MASTER}:${MASTER}/g" /etc/systemd/system/resilio-sync.service
     systemctl daemon-reload
     sed -i "s/BTSGUIP/$BTSYNCIP/g" /etc/resilio-sync/config.json
-    systemctl restart resilio-sync
+    if [[ $active == "active" ]]; then
+      systemctl start resilio-sync
+    fi
   fi
 fi
